Login Component Basic Format
HTML and CSS for the login component. Still need to make it a controlled component so we can "log the user in" - I'm thinking we set a browser cookie or use localStorage, since we can easily access it from anywhere in the application and render a different page depending on the value.